Conflicts:
	Community Balance Patch/Balance Changes/CoreDefines.sql
	CvGameCoreDLL_Expansion2/CvCity.cpp
	CvGameCoreDLL_Expansion2/CvDangerPlots.cpp
	CvGameCoreDLL_Expansion2/CvEconomicAI.cpp
	CvGameCoreDLL_Expansion2/CvHomelandAI.cpp
	CvGameCoreDLL_Expansion2/CvMilitaryAI.cpp
	CvGameCoreDLL_Expansion2/CvPlayerAI.cpp
	CvGameCoreDLL_Expansion2/CvTacticalAI.cpp
	CvGameCoreDLL_Expansion2/CvUnit.cpp

+++ CvGameCoreDLL_Expansion2/CvBarbarians.cpp
@@ -1040,8 +1040,11 @@ void CvBarbarians::DoSpawnBarbarianUnit(CvPlot* pPlot, bool bIgnoreMaxBarbarians
 
 #if defined(MOD_BUGFIX_MINOR)
 				// Stop units from plundered trade routes ending up in the ocean
-				if (!pUnit->jumpToNearestValidPlot())
-					pUnit->kill(false);	// Could not find a valid spot!
+				if(bIgnoreMaxBarbarians)
+				{
+					if (!pUnit->jumpToNearestValidPlot())
+						pUnit->kill(false);	// Could not find a valid spot!
+				}
 #endif
 			}
 		}
